old/requirements.txt
numpy==2.2.6
pillow==11.2.1
pygame==2.6.1
PyYAML==6.0.2
requests==2.32.4
urllib3==2.4.0
